feat: Move data retrieval to SensorServer (step-1)
Implement the SensorServer as a GenServer. During
initialization we invoke the processes to get
the data asyncronously and cache it on the server
state. In next step we will make the cache to be
refreshed at a fixed interval.

Also, as an exercise, we introduce into the SensorServer
initialization process the logic needed to make it
create random stamped snapshots or not given an
initialization value.

We used this last functionality to make the tests
predictable.