feat: Add KickStarter server to control HttpServer
As HttpServer is not a GenServer it can not be
supervisable by OTP supervisors. To manage this
we add a KickStarter module that acts as a GenServer
wrapper over HttpServer.

This new module spawns and links to a new HTTPServer
at start and creates a new one every time that module
dies.